Save form values to redux state with redux-form
