+++ app/containers/CheckRegForm/constants.js
@@ -1,0 +1,5 @@
+/* Field type constants */
+export const TEXT_FIELD = 'TEXT';
+export const SELECT_FIELD = 'SELECT';
+export const CHECK_FIELD = 'CHECK';
+export const HIDDEN_FIELD = 'HIDDEN';
