+++ app/containers/CheckRegForm/styles.css
@@ -1,0 +1,3 @@
+.checkRegForm {  /* stylelint-ignore */
+
+}
