+++ app/containers/CheckRegForm/tests/index.test.js
@@ -1,0 +1,11 @@
+// import CheckRegForm from '../index';
+
+import expect from 'expect';
+// import { shallow } from 'enzyme';
+// import React from 'react';
+
+describe('<CheckRegForm />', () => {
+  it('Expect to have unit tests specified', () => {
+    expect(true).toEqual(false);
+  });
+});
