+++ app/containers/CheckRegPage/constants.js
@@ -13,10 +13,3 @@ export const FETCH_STATES_URL = `${cfg.API_BASE}/check-registration-form/`;
 export const LOAD_STATES = 'app/CheckRegPage/LOAD_STATES';
 export const CHANGE_STATE = 'app/CheckRegPage/CHANGE_STATE';
 export const LOAD_STATE_FORM = 'app/CheckRegPage/LOAD_STATE_FORM';
-
-/* Field type constants */
-
-export const TEXT_FIELD = 'TEXT';
-export const SELECT_FIELD = 'SELECT';
-export const CHECK_FIELD = 'CHECK';
-export const HIDDEN_FIELD = 'HIDDEN';
