+++ app/containers/CheckRegPage/selectors.js
@@ -3,7 +3,7 @@ import { createSelector } from 'reselect';
 /**
  * Direct selector to the checkRegPage state domain
  */
-const selectCheckRegPageDomain = () => (state) => state.get('checkRegPage');
+const selectCheckRegPageDomain = () => (state) => state.get('CheckRegPage');
 
 const selectStates = () => createSelector(
   selectCheckRegPageDomain(),
