+++ app/routes.js
@@ -30,7 +30,7 @@ export default function createRoutes(store) {
         const renderRoute = loadModule(cb);
 
         importModules.then(([reducer, sagas, component]) => {
-          injectReducer('checkRegPage', reducer.default);
+          injectReducer('CheckRegPage', reducer.default);
           injectSagas(sagas.default);
 
           renderRoute(component);
