+++ package.json
@@ -61,7 +61,9 @@
     ],
     "env": {
       "production": {
-        "only": ["app"],
+        "only": [
+          "app"
+        ],
         "plugins": [
           "transform-react-remove-prop-types",
           "transform-react-constant-elements",
@@ -92,16 +94,31 @@
     },
     "rules": {
       "no-unused-vars": 1,
-      "space-before-function-paren": [2, "never"],
+      "space-before-function-paren": [
+        2,
+        "never"
+      ],
       "import/no-unresolved": 2,
-      "comma-dangle": [2, "always-multiline"],
-      "indent": [2, 2, { "SwitchCase": 1 }],
+      "comma-dangle": [
+        2,
+        "always-multiline"
+      ],
+      "indent": [
+        2,
+        2,
+        {
+          "SwitchCase": 1
+        }
+      ],
       "no-console": 1,
       "max-len": 0,
       "prefer-template": 2,
       "no-use-before-define": 0,
       "newline-per-chained-call": 0,
-      "arrow-body-style": [1, "as-needed"],
+      "arrow-body-style": [
+        1,
+        "as-needed"
+      ],
       "jsx-a11y/href-no-hash": 2,
       "jsx-a11y/label-has-for": 2,
       "jsx-a11y/mouse-events-have-key-events": 2,
@@ -124,9 +141,14 @@
       "color-hex-case": "upper",
       "string-quotes": "single",
       "font-family-name-quotes": "always-where-recommended",
-      "selector-pseudo-class-no-unknown": [true, {
-        "ignorePseudoClasses": ["global"]
-      }],
+      "selector-pseudo-class-no-unknown": [
+        true,
+        {
+          "ignorePseudoClasses": [
+            "global"
+          ]
+        }
+      ],
       "indentation": 2
     }
   },
@@ -158,15 +180,16 @@
     "intl": "1.2.4",
     "ip": "1.1.3",
     "minimist": "1.2.0",
-    "react": "15.2.1",
+    "react": "15.3.0",
     "react-dom": "15.2.1",
     "react-helmet": "3.1.0",
-    "react-redux": "4.4.5",
     "react-intl": "^2.1.3",
+    "react-redux": "4.4.5",
     "react-router": "2.6.0",
     "react-router-redux": "4.0.5",
     "react-router-scroll": "0.2.1",
     "redux": "3.5.2",
+    "redux-form": "6.0.0-rc.3",
     "redux-immutable": "3.0.6",
     "redux-saga": "0.11.0",
     "reselect": "2.5.3",
